requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
langchain-openai
sentence-transformers
faiss-cpu
pydantic>=2
python-dotenv
numpy
scikit-learn
httpx
tenacity
tqdm
langchain-huggingface
sentence-transformers
transformers
langchain-ollama
mlflow